fix android kotlin gradle

+++ android/app/build.gradle.kts
@@ -51,10 +51,10 @@ android {
     multiDexEnabled = true
 
     versionCode =
-        flutter.versionCode
+    flutter.versionCode.toInt()
 
-    versionName =
-        flutter.versionName
+versionName =
+    flutter.versionName
 }
 
     buildTypes {
